+++ graal/com.oracle.graal.hotspot/src/com/oracle/graal/hotspot/CompileTheWorld.java
@@ -279,7 +279,11 @@ public final class CompileTheWorld {
     }
 
     public void println(String s) {
-        if (verbose) {
+        println(verbose, s);
+    }
+
+    public static void println(boolean cond, String s) {
+        if (cond) {
             TTY.println(s);
         }
     }
@@ -605,6 +609,12 @@ public final class CompileTheWorld {
                                     compileMethod(javaMethod);
                                 }
                             }
+
+                            // Also compile the class initializer if it exists
+                            HotSpotResolvedJavaMethod clinit = (HotSpotResolvedJavaMethod) metaAccess.lookupJavaType(javaClass).getClassInitializer();
+                            if (clinit != null && canBeCompiled(clinit, clinit.getModifiers())) {
+                                compileMethod(clinit);
+                            }
                         }
                     } catch (Throwable t) {
                         println("CompileTheWorld (%d) : Skipping %s %s", classFileCounter, className, t.toString());
@@ -714,12 +724,16 @@ public final class CompileTheWorld {
      *
      * @return true if it can be compiled, false otherwise
      */
-    private static boolean canBeCompiled(HotSpotResolvedJavaMethod javaMethod, int modifiers) {
+    private boolean canBeCompiled(HotSpotResolvedJavaMethod javaMethod, int modifiers) {
         if (Modifier.isAbstract(modifiers) || Modifier.isNative(modifiers)) {
             return false;
         }
         HotSpotVMConfig c = config();
         if (c.dontCompileHugeMethods && javaMethod.getCodeSize() > c.hugeMethodLimit) {
+            println(verbose || methodFilters != null,
+                            String.format("CompileTheWorld (%d) : Skipping huge method %s (use -XX:-DontCompileHugeMethods or -XX:HugeMethodLimit=%d to include it)", classFileCounter,
+                                            javaMethod.format("%H.%n(%p):%r"),
+                                            javaMethod.getCodeSize()));
             return false;
         }
         // Allow use of -XX:CompileCommand=dontinline to exclude problematic methods